# docs/openapi/repository-api-v1.json
{
  "components": {
    "schemas": {
...
      "Body_publishSkillVersion": {
        "properties": {
          "artifact": {
            "contentMediaType": "application/octet-stream",
            "description": "Binary artifact associated with the published version.",
            "title": "Artifact",
            "type": "string"
          },
          "manifest": {
            "description": "Manifest JSON string describing the immutable skill version.",
            "title": "Manifest",
            "type": "string"
          }
        },
        "required": [
          "manifest",
          "artifact"
        ],
        "title": "Body_publishSkillVersion",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "description": "Standardized error envelope shared across HTTP endpoints.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Normalized error payload."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ExactSkillVersionResponse": {
        "description": "Exact immutable version metadata returned by the new fetch APIs.",
        "properties": {
          "artifact_ref": {
            "$ref": "#/components/schemas/ArtifactRefResponse",
            "description": "Backend-agnostic reference used to stream the immutable artifact."
          },
          "checksum": {
            "$ref": "#/components/schemas/ChecksumResponse",
            "description": "Checksum metadata for integrity checks."
          },
          "manifest": {
            "$ref": "#/components/schemas/SkillManifest",
            "description": "Validated manifest as stored by the service."
          },
          "published_at": {
            "description": "UTC timestamp when the version was published.",
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "skill_id": {
            "description": "Stable identifier of the skill.",
            "title": "Skill Id",
            "type": "string"
          },
          "version": {
            "description": "Immutable semantic version of the skill.",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "skill_id",
          "version",
          "manifest",
          "checksum",
          "artifact_ref",
          "published_at"
        ],
        "title": "ExactSkillVersionResponse",
        "type": "object"
      },
...
      "SkillVersionListResponse": {
        "description": "List of all known immutable versions for a skill.",
        "properties": {
          "skill_id": {
            "description": "Stable identifier of the skill.",
            "title": "Skill Id",
            "type": "string"
          },
          "versions": {
            "description": "Published versions in deterministic reverse chronological order.",
            "items": {
              "$ref": "#/components/schemas/SkillVersionSummaryResponse"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "skill_id",
          "versions"
        ],
        "title": "SkillVersionListResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Registry-first API for immutable skill publication, indexed search candidate\nretrieval, direct relationship reads, exact version retrieval, artifact streaming,\nand version listing.\n\nThe server owns data-local registry operations only. Prompt interpretation,\nreranking, dependency solving, final selection, lock generation, and execution\nplanning remain client-owned behavior outside this API boundary.",
    "title": "aptitude-server Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/fetch/skills/{skill_id}/{version}": {
      "get": {
        "description": "Return the stored manifest, checksum, and a backend-agnostic artifact reference for one exact immutable skill version. This route does not inline artifact bytes.",
        "operationId": "getExactSkillVersionMetadata",
        "parameters": [
          {
            "description": "Stable identifier of the skill to fetch.",
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "description": "Stable identifier of the skill to fetch.",
              "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
              "title": "Skill Id",
              "type": "string"
            }
          },
          {
            "description": "Exact immutable semantic version to fetch.",
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "description": "Exact immutable semantic version to fetch.",
              "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "artifact_ref": {
                    "checksum_algorithm": "sha256",
                    "checksum_digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2",
                    "download_path": "/fetch/skills/python.lint/1.2.3/artifact",
                    "size_bytes": 123
                  },
                  "checksum": {
                    "algorithm": "sha256",
                    "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                  },
                  "manifest": {
                    "conflicts_with": [],
                    "depends_on": [
                      {
                        "markers": [
                          "linux",
                          "gpu"
                        ],
                        "optional": true,
                        "skill_id": "python.base",
                        "version_constraint": ">=1.0.0,<2.0.0"
                      }
                    ],
                    "description": "Linting skill",
                    "extends": [
                      {
                        "skill_id": "python.base",
                        "version": "1.0.0"
                      }
                    ],
                    "name": "Python Lint",
                    "overlaps_with": [],
                    "schema_version": "1.0",
                    "skill_id": "python.lint",
                    "tags": [
                      "python",
                      "lint"
                    ],
                    "version": "1.2.3"
                  },
                  "published_at": "2026-03-10T08:30:00Z",
                  "skill_id": "python.lint",
                  "version": "1.2.3"
                },
                "schema": {
                  "$ref": "#/components/schemas/ExactSkillVersionResponse"
                }
              }
            },
            "description": "Exact immutable skill version metadata returned successfully."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SKILL_VERSION_NOT_FOUND",
                    "details": {
                      "skill_id": "python.lint",
                      "version": "9.9.9"
                    },
                    "message": "Skill version not found: python.lint@9.9.9"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested immutable `skill_id@version` does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body or path parameters are invalid."
          }
        },
        "summary": "Fetch exact immutable version metadata",
        "tags": [
          "fetch"
        ]
      }
    },
...
    "/skills/publish": {
      "post": {
        "description": "Accept multipart form data containing a JSON `manifest` string and a binary `artifact` upload. The manifest is validated at the API boundary, converted to core domain models, and persisted only if the same `skill_id@version` does not already exist.",
        "operationId": "publishSkillVersion",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "examples": {
                "publish": {
                  "summary": "Publish a new immutable skill version",
                  "value": {
                    "artifact": "(binary artifact upload)",
                    "manifest": "{\"schema_version\": \"1.0\", \"skill_id\": \"python.lint\", \"version\": \"1.2.3\", \"name\": \"Python Lint\", \"description\": \"Linting skill\", \"tags\": [\"python\", \"lint\"], \"depends_on\": [{\"skill_id\": \"python.base\", \"version_constraint\": \">=1.0.0,<2.0.0\", \"optional\": true, \"markers\": [\"linux\", \"gpu\"]}], \"extends\": [{\"skill_id\": \"python.base\", \"version\": \"1.0.0\"}], \"conflicts_with\": [], \"overlaps_with\": []}"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/Body_publishSkillVersion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "artifact_metadata": {
                    "relative_path": "python.lint/1.2.3/artifact.bin",
                    "size_bytes": 123
                  },
                  "checksum": {
                    "algorithm": "sha256",
                    "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                  },
                  "manifest": {
                    "conflicts_with": [],
                    "depends_on": [
                      {
                        "markers": [
                          "linux",
                          "gpu"
                        ],
                        "optional": true,
                        "skill_id": "python.base",
                        "version_constraint": ">=1.0.0,<2.0.0"
                      }
                    ],
                    "description": "Linting skill",
                    "extends": [
                      {
                        "skill_id": "python.base",
                        "version": "1.0.0"
                      }
                    ],
                    "name": "Python Lint",
                    "overlaps_with": [],
                    "schema_version": "1.0",
                    "skill_id": "python.lint",
                    "tags": [
                      "python",
                      "lint"
                    ],
                    "version": "1.2.3"
                  },
                  "published_at": "2026-03-10T08:30:00Z",
                  "skill_id": "python.lint",
                  "version": "1.2.3"
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillVersionDetailResponse"
                }
              }
            },
            "description": "Immutable skill version published successfully."
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "DUPLICATE_SKILL_VERSION",
                    "details": {
                      "skill_id": "python.lint",
                      "version": "1.2.3"
                    },
                    "message": "Skill version already exists: python.lint@1.2.3"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested immutable `skill_id@version` already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "invalid_manifest": {
                    "value": {
                      "error": {
                        "code": "INVALID_MANIFEST",
                        "details": {
                          "errors": [
                            {
                              "input": "not-json",
                              "loc": [],
                              "msg": "Invalid JSON: expected value at line 1 column 1",
                              "type": "json_invalid"
                            }
                          ]
                        },
                        "message": "Manifest validation failed."
                      }
                    }
                  },
                  "invalid_request": {
                    "value": {
                      "error": {
                        "code": "INVALID_REQUEST",
                        "details": {
                          "errors": [
                            {
                              "input": "latest",
                              "loc": [
                                "path",
                                "version"
                              ],
                              "msg": "String should match pattern",
                              "type": "string_pattern_mismatch"
                            }
                          ]
                        },
                        "message": "Request validation failed."
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request is invalid or the manifest JSON violates the publish contract."
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "ARTIFACT_STORAGE_FAILURE",
                    "message": "Artifact storage failed during publish."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Artifact persistence failed while publishing the version."
          }
        },
        "summary": "Publish an immutable skill version",
        "tags": [
          "skills"
        ]
      }
    },
...
    "/skills/{skill_id}": {
      "get": {
        "description": "Return every published immutable version for `skill_id`. Results are deterministic and ordered newest-first by publication timestamp.",
        "operationId": "listSkillVersions",
        "parameters": [
          {
            "description": "Stable identifier of the skill whose versions should be listed.",
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "description": "Stable identifier of the skill whose versions should be listed.",
              "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
              "title": "Skill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "skill_id": "python.lint",
                  "versions": [
                    {
                      "artifact_metadata": {
                        "relative_path": "python.lint/1.2.3/artifact.bin",
                        "size_bytes": 123
                      },
                      "checksum": {
                        "algorithm": "sha256",
                        "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                      },
                      "manifest": {
                        "conflicts_with": [],
                        "depends_on": [
                          {
                            "markers": [
                              "linux",
                              "gpu"
                            ],
                            "optional": true,
                            "skill_id": "python.base",
                            "version_constraint": ">=1.0.0,<2.0.0"
                          }
                        ],
                        "description": "Linting skill",
                        "extends": [
                          {
                            "skill_id": "python.base",
                            "version": "1.0.0"
                          }
                        ],
                        "name": "Python Lint",
                        "overlaps_with": [],
                        "schema_version": "1.0",
                        "skill_id": "python.lint",
                        "tags": [
                          "python",
                          "lint"
                        ],
                        "version": "1.2.3"
                      },
                      "published_at": "2026-03-10T08:30:00Z",
                      "skill_id": "python.lint",
                      "version": "1.2.3"
                    },
                    {
                      "artifact_metadata": {
                        "relative_path": "python.lint/1.0.0/artifact.bin",
                        "size_bytes": 98
                      },
                      "checksum": {
                        "algorithm": "sha256",
                        "digest": "2d711642b726b04401627ca9fbac32f5da7e5a3d4f2e9ce1d49f6b2e5f3b1f9d"
                      },
                      "manifest": {
                        "conflicts_with": [],
                        "depends_on": [],
                        "description": "Linting skill",
                        "extends": [],
                        "name": "Python Lint",
                        "overlaps_with": [],
                        "schema_version": "1.0",
                        "skill_id": "python.lint",
                        "tags": [
                          "python",
                          "lint"
                        ],
                        "version": "1.0.0"
                      },
                      "published_at": "2026-02-10T08:30:00Z",
                      "skill_id": "python.lint",
                      "version": "1.0.0"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillVersionListResponse"
                }
              }
            },
            "description": "Published versions listed successfully."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request payload, form fields, path parameters, or query parameters are invalid."
          }
        },
        "summary": "List published versions for a skill",
        "tags": [
          "skills"
        ]
      }
    },
...
  }
}